refactor: Store monetary values as integers
* Refactor: Store monetary values as integers

This commit changes the storage of monetary values (user balances,
order accruals, and withdrawals) from floating-point numbers (`real`)
to integers in the database. All monetary values are now stored as
cents (e.g., $205.60 is stored as 20560).

Key changes include:
- Database migration to change column types to `integer` and convert existing data.
- Updates to `sqlc` schema, queries, and generated models.
- Application models updated to use `int32` for monetary fields internally.
- Storage layer (`internal/postgres/storage.go`) updated to:
    - Convert incoming `float32` DTO values to `int32` cents before DB storage.
    - Map `int32` cents from DB to `int32` in application models.
- Service layer (`internal/application/service/service.go`) updated to:
    - Receive `int32` cents from the storage layer.
    - Convert these `int32` cents to `float32` decimal values for API responses,
      ensuring the external API contract remains unchanged.
- All relevant tests (service layer, job tests) updated to reflect these
  changes, including mock storage interactions and assertions.

This change mitigates potential floating-point precision issues and
ensures more accurate monetary calculations and storage.

* fix: fix some logic

* style: remove comments

---------

+++ internal/application/model/order.go
@@ -11,10 +11,10 @@ type OrderStatus string
 
 // List of possible order statuses
 const (
-	OrderStatusNew       OrderStatus = "NEW"
+	OrderStatusNew        OrderStatus = "NEW"
 	OrderStatusProcessing OrderStatus = "PROCESSING"
-	OrderStatusInvalid   OrderStatus = "INVALID"
-	OrderStatusProcessed OrderStatus = "PROCESSED"
+	OrderStatusInvalid    OrderStatus = "INVALID"
+	OrderStatusProcessed  OrderStatus = "PROCESSED"
 )
 
 // AccrualOrderStatus represents the status of an order in the accrual system
@@ -34,7 +34,7 @@ type Order struct {
 	UserID    uuid.UUID
 	CreatedAt time.Time
 	Number    string
-	Accrual   float32
+	Accrual   int32
 	Status    OrderStatus
 }
